Moved CFBundle.Architecture to CourFoundation Core Sources
